flask backend (save for cors) done

+++ resources/recipes.py
@@ -4,6 +4,8 @@ from flask import Blueprint, request, jsonify
 
 from playhouse.shortcuts import model_to_dict
 
+from flask_login import current_user, login_required
+
 recipes = Blueprint('recipes', 'recipes')
 
 @recipes.route('/', methods=['GET'])
@@ -23,12 +25,13 @@ def recipes_index():
 
 
 @recipes.route('/', methods=['POST'])
+@login_required
 def create_recipe():
     payload = request.get_json()
 
     new_recipe = models.Recipe.create(
         name=payload['name'],
-        poster=payload['poster'],
+        poster=current_user.id,
         origin=payload['origin'],
         ingredients=payload['ingredients'],
         instructions=payload['instructions'],
@@ -36,7 +39,7 @@ def create_recipe():
 
     recipe_dict = model_to_dict(new_recipe)
 
-    print(recipe_dict)
+    recipe_dict['poster'].pop('password')
 
     return jsonify(
         data=recipe_dict,
@@ -48,43 +51,75 @@ def create_recipe():
 # recipe destroy route
 
 @recipes.route('/<id>', methods=['DELETE'])
+@login_required
 def delete_recipe(id):
+
+    try:
         recipe_to_delete = models.Recipe.get_by_id(id)
 
-        recipe_to_delete.delete_instance()
+        if recipe_to_delete.poster.id == current_user.id:
 
+            recipe_to_delete.delete_instance()
+
+            return jsonify(
+                data={},
+                message="Successfully deleted recipe with the id of {}".format(id),
+                status=200
+            ), 200
+
+        else:
+            return jsonify(
+                data={},
+                message="Not your recipe!",
+                status=403
+            ), 403
+
+    except models.DoesNotExist:
         return jsonify(
             data={},
-            message="Successfully deleted recipe with the id of {}".format(id),
-            status=200
-        ), 200
+            message="There's no recipe with that ID",
+            status=404
+        ), 404
+
 
 
 # update route
 @recipes.route('/<id>', methods=['PUT'])
+@login_required
 def update_recipe(id):
     payload = request.get_json()
 
     recipe_to_update = models.Recipe.get_by_id(id)
 
-    if 'name' in payload:
-        recipe_to_update.name = payload['name']
-    if 'origin' in payload:
-        recipe_to_update.origin = payload['origin']
-    if 'ingredients' in payload:
-        recipe_to_update.ingredients = payload['ingredients']
-    if 'instructions' in payload:
-        recipe_to_update.instructions = payload['instructions']
+    if recipe_to_update.poster.id == current_user.id:
 
-    recipe_to_update.save()
+        if 'name' in payload:
+            recipe_to_update.name = payload['name']
+        if 'origin' in payload:
+            recipe_to_update.origin = payload['origin']
+        if 'ingredients' in payload:
+            recipe_to_update.ingredients = payload['ingredients']
+        if 'instructions' in payload:
+            recipe_to_update.instructions = payload['instructions']
 
-    updated_recipe_dict = model_to_dict(recipe_to_update)
+        recipe_to_update.save()
 
-    return jsonify(
-        data=updated_recipe_dict,
-        message="Successfully updated the recipe with id of {}".format(id),
-        status = 200
-    ), 200
+        updated_recipe_dict = model_to_dict(recipe_to_update)
+
+        updated_recipe_dict['poster'].pop('password')
+
+        return jsonify(
+            data=updated_recipe_dict,
+            message="Successfully updated the recipe with id of {}".format(id),
+            status = 200
+        ), 200
+
+    else:
+        return jsonify(
+            data={},
+            message="Not your recipe!!",
+            status=403
+        ), 403
 
 @recipes.route('/<id>', methods=['GET'])
 def show_recipe(id):
